backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
pytz==2024.1
prometheus-flask-exporter==0.23.2
prometheus-client==0.19.0
opentelemetry-api==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-instrumentation-flask==0.48b0
requests==2.31.0
opentelemetry-instrumentation-requests==0.48b0
opentelemetry-exporter-otlp-proto-grpc==1.27.0